Camera Look Rotation and Player Rotation
Implemented a Player look script that handles camera rotation on the X axis. It stores a body rotation that the Player Controller script uses to rotate itself on the Up axis. Since the camera is childed to the player, the camera in turn will follow the rotation. Need to implement falling and jumping.